backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.11
alembic>=1.13.3
pydantic>=2.9.2
pydantic-settings>=2.5.2
pydantic[email]>=2.9.2
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
yfinance>=0.2.43
curl_cffi>=0.7.4
pandas>=2.2.0
requests>=2.32.3
pdfplumber>=0.11.4
scipy>=1.13.0
numpy>=1.26.0
pytest>=7.4.0
pytest-cov>=4.0.0
httpx>=0.27.0
